+++ vms/platformvm/utxo/handler.go
@@ -199,7 +199,7 @@ func (h *handler) Spend(
 			// iteration of the UTXO set
 			continue
 		}
-		if out.Locktime <= now {
+		if out.Locktime <= uint64(now) {
 			// This output is no longer locked, so it will be handled during the
 			// next iteration of the UTXO set
 			continue
@@ -295,7 +295,7 @@ func (h *handler) Spend(
 		out := utxo.Out
 		inner, ok := out.(*stakeable.LockOut)
 		if ok {
-			if inner.Locktime > now {
+			if inner.Locktime > uint64(now) {
 				// This output is currently locked, so this output can't be
 				// burned. Additionally, it may have already been consumed
 				// above. Regardless, we skip to the next UTXO
@@ -484,7 +484,7 @@ func (h *handler) VerifySpendUTXOs(
 	}
 
 	// Time this transaction is being verified
-	now := h.clk.Time().Unix()
+	now := uint64(h.clk.Time().Unix())
 
 	// Track the amount of unlocked transfers
 	// assetID -> amount
@@ -492,8 +492,8 @@ func (h *handler) VerifySpendUTXOs(
 
 	// Track the amount of locked transfers and their owners
 	// assetID -> locktime -> ownerID -> amount
-	lockedProduced := make(map[ids.ID]map[int64]map[ids.ID]uint64)
-	lockedConsumed := make(map[ids.ID]map[int64]map[ids.ID]uint64)
+	lockedProduced := make(map[ids.ID]map[uint64]map[ids.ID]uint64)
+	lockedConsumed := make(map[ids.ID]map[uint64]map[ids.ID]uint64)
 
 	for index, input := range ins {
 		utxo := utxos[index] // The UTXO consumed by [input]
@@ -510,7 +510,7 @@ func (h *handler) VerifySpendUTXOs(
 		}
 
 		out := utxo.Out
-		locktime := int64(0)
+		locktime := uint64(0)
 		// Set [locktime] to this UTXO's locktime, if applicable
 		if inner, ok := out.(*stakeable.LockOut); ok {
 			out = inner.TransferableOut
@@ -563,7 +563,7 @@ func (h *handler) VerifySpendUTXOs(
 		}
 		lockedConsumedAsset, ok := lockedConsumed[realAssetID]
 		if !ok {
-			lockedConsumedAsset = make(map[int64]map[ids.ID]uint64)
+			lockedConsumedAsset = make(map[uint64]map[ids.ID]uint64)
 			lockedConsumed[realAssetID] = lockedConsumedAsset
 		}
 		ownerID := hashing.ComputeHash256Array(ownerBytes)
@@ -583,7 +583,7 @@ func (h *handler) VerifySpendUTXOs(
 		assetID := out.AssetID()
 
 		output := out.Output()
-		locktime := int64(0)
+		locktime := uint64(0)
 		// Set [locktime] to this output's locktime, if applicable
 		if inner, ok := output.(*stakeable.LockOut); ok {
 			output = inner.TransferableOut
@@ -612,7 +612,7 @@ func (h *handler) VerifySpendUTXOs(
 		}
 		lockedProducedAsset, ok := lockedProduced[assetID]
 		if !ok {
-			lockedProducedAsset = make(map[int64]map[ids.ID]uint64)
+			lockedProducedAsset = make(map[uint64]map[ids.ID]uint64)
 			lockedProduced[assetID] = lockedProducedAsset
 		}
 		ownerID := hashing.ComputeHash256Array(ownerBytes)
